Bug #43106 Falcon recovery assertion (bdb->buffer->pageNumber == pageNumber) in Cache.cpp
The problem is that Falcon during recovery can access a data page that can 
be invalid. If the same transaction creates, and deletes the same blob but
does not finish before the crash, data page s not guaranteed to be written to disk
before delete.

The fix is to update data page  only if transaction has completed, and we
that data page is on disk. Blob update was already handled this way.