Bug#44040 MySQL allows creating a MERGE table upon VIEWs but          crashes when using it.
In 6.0 the crash occurred in myisammrg_attach_children_callback
while checking the table version for compatibility.

Due to code changes the fix in 6.0 is in method open_table but fix
is same. A check introduced to check if view is a child table of
a merge table and returned error if it is.