BUG#40045: Replication failure on RBR + no PK + UPDATE + integer key + char with no key
Row based replication would break if updating a row and using a key
for searching for the correct row to update. This was observed when
using MyISAM storage engine on the slave. Since MyISAM does not
provide partial row fetch (HA_PARTIAL_COLUMN_READ) , the comparison
between the row fetched on the slave and the before image (BI)
replicated from master would fail. This happened because the BI (which
sometimes was only part of the row) would always be compared against a
complete row.
                  
Furthermore, the read set mask was not being properly set in some cases
when using keys (not unique nor primary) which would lead to partial 
row matches, thence changes might end up in the wrong row.
                  
This patch addresses this by changing the record_compare function for
taking into consideration the fields set on the readset and the engine
used on the slave. Furthermore, it also moves the mark columns needed
for update on the master so that read_set bitmap is always correct.