Fix for Bug#43138: DROP DATABASE failure does not clean up message list.
The problem was that the high-level function mysql_rm_db() invoked
low-level mysql_rm_table_part2(), which reported low-level error
(Unknown table) if SE refused to delete a table. Also when
mysql_rm_table_part2() reported an error, it didn't add corresponding
warning into the list (because it is used from other places where such
behaviour is required).

The fix is to
  1. Remove no_warnings_for_error usage from sql_table.cc
  2. Improve internal error handler support in THD, so that
     a stack of error handlers is allowed.
  3. Create an internal error handler (Drop_table_error_handler)
     to silence useless warnings.
  4. Use the handler in DROP DATABASE and DROP TABLE statements.